Handle outputing parsing error cases
Added the negative test file that causes the code to error incorrectly.

If the code lacks calls to BeginSourceFile / EndSourceFile while parsing or
lexing, the code will assert / fault because its expecting to handle
commandline errors, not file errors -- basically inconsistent state.

TODO: While investigating this, I noticed that tutorials 3 and 4 appear to
be pulling in implicit headers from somewhere -- I wonder where that is
occuring and why that is.

TODO: Need to make it easier to work with the different test cases from the
command line instead of having it hardcoded -- nice to have.