Update samples to build against 168294 in main.
Most of the interfaces that take options objects now use an llvm smart
pointer API. Further, in the cases where I'd initially written up a few
bits of code that use the LLVM smart points API years ago, I used them
incorrectly and was passing null values. The most recent changes in
Clang just turned that up so I've fixed some of the cases of that (there
are more that don't cause crashes).

Work still to do:
Fix the rest of the pointer usage

Fix the naming of variables and indentation to be clean and consistent.

I think in the refactoring, I may have left references to types floating
around that I shouldn't have.