Update the readme based on feedback.
Beatrice Creusillet pointed out that I was making an assumption about
how users installed clang and llvm without making that explicit.

Adding that assumption to the readme which is that users have installed
clang and llvm on their system with "make install".

In the long term, it would probably be nice to have some configuration
setup across the various OSes that this project runs on.